requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
aiohttp
python-dotenv
numpy
fastapi-cors
bcrypt
sqlalchemy
python-jose[cryptography]
python-multipart
passlib[bcrypt]